Imported Sailor Moon Hairstyle from Goon2016
I didn't really want to do it this way - but I got it working; I had every intention of doing this in a seperate file - but because of the way sprite_accessories are coded I cannot just place in a hairstyle with a custom icon outside the default file setup. I really wish I could've gotten it working with all the goonstation 2016 resources in the /goon/icons files, but my understanding of the code just isn't there yet.