Changed Base Silicon Laws to Reflect Asimov 2.0
Asimov 2.0 is based off Asimov but with more leeway given to the machine in the terms of playability. These laws were HIGHLY influenced by Goonstation's Modified Asimov Lawset, and in particular it changes the lawset to reflect that inaction is no longer a volition of the robotic laws. This was a change brought about by the idea that an Robot should allow a person to commit suicide and not have to suffer a law breakdown.

Also Added Law 4: To Resolve Chain of Command Issues between different ranking members of the station.