Started Writing Lavaland Z-level Expansion
First Fully Functional Commit of the Z-level idea I had on projects. It is incomplete but I've successfully "Dug to map level 2". In addition to the map changes I've added a way to transcend them, at round start various up/down holes will be spawned on each level the holes lead up or down one map level (right now there's only 2 levels), when someone first attempts to traverse them they select an linking hole on another map level - for the rest of that round that hole will link to that exit; however the hole on the other end may not lead "up" to the same exit on the other side. Down Holes you can fall through; Up Holes you need a rope to traverse by "attacking" the hole with the rope while on it's tile (similiar to placing a holocall). Also did a second pass on air transferrence on te sleepsack as a side-project. Finally you can obtain ropes from the new "Public" Mining Vendors (they are actual vendors so they have stock unlike the actual mining vendor) as well as a single rope spawn that's on every map level below the first (currently only map level 2). Difficulty is not yet a factor in changing map levels, however map level two is essentially "underneath" the mining outpost of lavaland and shares a largely similiar map design and has the same (at the moment) random terrain generation.