A Few Tweaks
First and foremost - I changed the default assistant wear to be gender-based; also I added the holiday "Girly Dev Day" in honor of well...me. The Holiday will trigger every March 2nd (happens to be my rl bday) and will dress all the assistants in lockable schoolgirl uniforms. Additionally I added a "Pink Dress" which is not yet available outside admin spawn (will be added to the vendors soon I just haven't gotten around to it yet.) Finally To Address Rumors that the fork is dead - it's not I just decided to take a mild break from game development.