Silicons can now Vend for Humans
Little Goon-inspired feature: silicons like AI and Borgs can now "Hack" a vendor to make it vend an item for a human. Note this does not allow them to vend endlessly because the item count is still limited by stock; they do however not have to pay for the item. Silicons can only use this ability for vending machines on the station not on Z-Maps.

- no_silicons variable added to Vendors; When TRUE prevents the silicons from accessing the item at all